backend/requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Libraries - Use compatible versions
langchain==0.0.350
langgraph==0.0.20
openai==1.3.0
anthropic==0.8.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Calendar Integration
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

# Frontend
streamlit==1.28.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.5.0
python-dateutil==2.8.2
pytz==2023.3

# Additional dependencies for modern functionality
aioredis==2.0.1
httpx==0.27.0

# Core dependencies - use pre-compiled wheels
numpy
pandas
